knn algorithm

+++ src/minsk/tree/Pair.java
@@ -2,6 +2,6 @@ package minsk.tree;
 
 
 public class Pair {
-	public int l;
-	public int h;
+	public double l;
+	public double h;
 }
